Adding feeback loop to ASO, add preserveLFN option
This adds the proper information to user_monitoring DB to let WMAgent
follow the status of ASO with the ASOTracker component (seperate patch)

For the perserveLFN option:

Currently ASO, blindly replaces 'store/temp' with 'store', but that's
not always what you want to do (consider, for instance, unmerged files).
This adds an optional flag 'preserveLFN' that can be set per-file that
declares, "this file shouldn't have its LFN changed". By default the old
behavior still happens